Tanks v0.7.0 - Added 3D Graphics and Party mode, and made many other changes.

+++ src/main/java/tanks/tank/TankWhite.java
@@ -1,8 +1,8 @@
 package tanks.tank;
 
+import tanks.Drawing;
 import tanks.Effect;
 import tanks.Game;
-import tanks.Drawing;
 
 public class TankWhite extends TankAIControlled
 {
@@ -33,7 +33,7 @@ public class TankWhite extends TankAIControlled
 					for (int i = 0; i < 50; i++)
 					{
 						Effect e = Effect.createNewEffect(this.posX, this.posY, Effect.EffectType.piece);
-						int var = 50;
+						double var = 50;
 						e.colR = Math.min(255, Math.max(0, this.colorR + Math.random() * var - var / 2));
 						e.colG = Math.min(255, Math.max(0, this.colorR + Math.random() * var - var / 2));
 						e.colB = Math.min(255, Math.max(0, this.colorR + Math.random() * var - var / 2));
